The hardware accelerator will not output the eventslice again, instead it will put a simplified event stream so it can be easily processed in the host PC. Due to UDP doesn't support framgementation, the output length of eventSlice is limited to 43200. It can render the binary event-slice on host PC in real time with the hardware accelerator. In the future, TCP should be tried again to make the eventslice size not fixed.

+++ mz_libcaer_OF/src/ABMOF/abmof_hw_accel.h
@@ -14,6 +14,6 @@
 #define DVS_WIDTH  240
 #define DVS_HEIGHT 180
 
-void parseEvents(const uint32_t * data, int32_t eventsArraySize, int8_t *eventSlice);
+void parseEvents(const uint32_t * data, int32_t eventsArraySize, int32_t *eventSlice);
 
 #endif
